feat: deprecate classic features removed by the scikit-build-core backend swap
The next major release (#1185) replaces the classic backend with
scikit-build-core's setuptools plugin. Warn ahead of that flip via a new
SKBuildDeprecationWarning(FutureWarning):

- setup() kwargs: cmake_with_sdist / non-default cmake_install_target
  (become errors) and cmake_languages / cmake_minimum_required_version
  (ignored) warn when set to a non-default value.
- Internal modules cmaker, constants, utils, command, platform_specifics,
  and setuptools_wrap warn on import. Each real module moved to a private
  _X; the public name is a shim (from ._X import * + warn). Internal imports
  and __init__ are repointed to _X so `from skbuild import setup` stays
  silent.

The suite globally ignores the warning (it imports these internals
pervasively); tests/test_deprecations.py asserts each warning explicitly.
Docs autodoc repointed to the private modules; removed a dead
skbuild.command.test reference left over from #1014.
