+++ docs/skbuild.command.rst
@@ -1,7 +1,7 @@
 skbuild.command package
 =======================
 
-.. automodule:: skbuild.command
+.. automodule:: skbuild._command
    :members:
    :undoc-members:
    :show-inheritance:
@@ -12,7 +12,7 @@ Submodules
 skbuild.command.bdist module
 ----------------------------
 
-.. automodule:: skbuild.command.bdist
+.. automodule:: skbuild._command.bdist
    :members:
    :undoc-members:
    :show-inheritance:
@@ -20,7 +20,7 @@ skbuild.command.bdist module
 skbuild.command.bdist\_wheel module
 -----------------------------------
 
-.. automodule:: skbuild.command.bdist_wheel
+.. automodule:: skbuild._command.bdist_wheel
    :members:
    :undoc-members:
    :show-inheritance:
@@ -28,7 +28,7 @@ skbuild.command.bdist\_wheel module
 skbuild.command.build module
 ----------------------------
 
-.. automodule:: skbuild.command.build
+.. automodule:: skbuild._command.build
    :members:
    :undoc-members:
    :show-inheritance:
@@ -36,7 +36,7 @@ skbuild.command.build module
 skbuild.command.build\_ext module
 ---------------------------------
 
-.. automodule:: skbuild.command.build_ext
+.. automodule:: skbuild._command.build_ext
    :members:
    :undoc-members:
    :show-inheritance:
@@ -44,7 +44,7 @@ skbuild.command.build\_ext module
 skbuild.command.build\_py module
 --------------------------------
 
-.. automodule:: skbuild.command.build_py
+.. automodule:: skbuild._command.build_py
    :members:
    :undoc-members:
    :show-inheritance:
@@ -52,7 +52,7 @@ skbuild.command.build\_py module
 skbuild.command.clean module
 ----------------------------
 
-.. automodule:: skbuild.command.clean
+.. automodule:: skbuild._command.clean
    :members:
    :undoc-members:
    :show-inheritance:
@@ -60,7 +60,7 @@ skbuild.command.clean module
 skbuild.command.egg\_info module
 --------------------------------
 
-.. automodule:: skbuild.command.egg_info
+.. automodule:: skbuild._command.egg_info
    :members:
    :undoc-members:
    :show-inheritance:
@@ -68,7 +68,7 @@ skbuild.command.egg\_info module
 skbuild.command.generate\_source\_manifest module
 -------------------------------------------------
 
-.. automodule:: skbuild.command.generate_source_manifest
+.. automodule:: skbuild._command.generate_source_manifest
    :members:
    :undoc-members:
    :show-inheritance:
@@ -76,7 +76,7 @@ skbuild.command.generate\_source\_manifest module
 skbuild.command.install module
 ------------------------------
 
-.. automodule:: skbuild.command.install
+.. automodule:: skbuild._command.install
    :members:
    :undoc-members:
    :show-inheritance:
@@ -84,7 +84,7 @@ skbuild.command.install module
 skbuild.command.install\_lib module
 -----------------------------------
 
-.. automodule:: skbuild.command.install_lib
+.. automodule:: skbuild._command.install_lib
    :members:
    :undoc-members:
    :show-inheritance:
@@ -92,7 +92,7 @@ skbuild.command.install\_lib module
 skbuild.command.install\_scripts module
 ---------------------------------------
 
-.. automodule:: skbuild.command.install_scripts
+.. automodule:: skbuild._command.install_scripts
    :members:
    :undoc-members:
    :show-inheritance:
@@ -100,7 +100,7 @@ skbuild.command.install\_scripts module
 skbuild.command.sdist module
 ----------------------------
 
-.. automodule:: skbuild.command.sdist
+.. automodule:: skbuild._command.sdist
    :members:
    :undoc-members:
    :show-inheritance:
