+++ docs/skbuild.platform_specifics.rst
@@ -1,7 +1,7 @@
 skbuild.platform\_specifics package
 ===================================
 
-.. automodule:: skbuild.platform_specifics
+.. automodule:: skbuild._platform_specifics
    :members:
    :undoc-members:
    :show-inheritance:
@@ -12,7 +12,7 @@ Submodules
 skbuild.platform\_specifics.abstract module
 -------------------------------------------
 
-.. automodule:: skbuild.platform_specifics.abstract
+.. automodule:: skbuild._platform_specifics.abstract
    :members:
    :undoc-members:
    :show-inheritance:
@@ -20,7 +20,7 @@ skbuild.platform\_specifics.abstract module
 skbuild.platform\_specifics.bsd module
 --------------------------------------
 
-.. automodule:: skbuild.platform_specifics.bsd
+.. automodule:: skbuild._platform_specifics.bsd
    :members:
    :undoc-members:
    :show-inheritance:
@@ -28,7 +28,7 @@ skbuild.platform\_specifics.bsd module
 skbuild.platform\_specifics.cygwin module
 -----------------------------------------
 
-.. automodule:: skbuild.platform_specifics.cygwin
+.. automodule:: skbuild._platform_specifics.cygwin
    :members:
    :undoc-members:
    :show-inheritance:
@@ -36,7 +36,7 @@ skbuild.platform\_specifics.cygwin module
 skbuild.platform\_specifics.linux module
 ----------------------------------------
 
-.. automodule:: skbuild.platform_specifics.linux
+.. automodule:: skbuild._platform_specifics.linux
    :members:
    :undoc-members:
    :show-inheritance:
@@ -44,7 +44,7 @@ skbuild.platform\_specifics.linux module
 skbuild.platform\_specifics.osx module
 --------------------------------------
 
-.. automodule:: skbuild.platform_specifics.osx
+.. automodule:: skbuild._platform_specifics.osx
    :members:
    :undoc-members:
    :show-inheritance:
@@ -52,7 +52,7 @@ skbuild.platform\_specifics.osx module
 skbuild.platform\_specifics.aix module
 ----------------------------------------
 
-.. automodule:: skbuild.platform_specifics.aix
+.. automodule:: skbuild._platform_specifics.aix
    :members:
    :undoc-members:
    :show-inheritance:
@@ -60,7 +60,7 @@ skbuild.platform\_specifics.aix module
 skbuild.platform\_specifics.platform\_factory module
 ----------------------------------------------------
 
-.. automodule:: skbuild.platform_specifics.platform_factory
+.. automodule:: skbuild._platform_specifics.platform_factory
    :members:
    :undoc-members:
    :show-inheritance:
@@ -68,7 +68,7 @@ skbuild.platform\_specifics.platform\_factory module
 skbuild.platform\_specifics.unix module
 ---------------------------------------
 
-.. automodule:: skbuild.platform_specifics.unix
+.. automodule:: skbuild._platform_specifics.unix
    :members:
    :undoc-members:
    :show-inheritance:
@@ -76,7 +76,7 @@ skbuild.platform\_specifics.unix module
 skbuild.platform\_specifics.windows module
 ------------------------------------------
 
-.. automodule:: skbuild.platform_specifics.windows
+.. automodule:: skbuild._platform_specifics.windows
    :members:
    :undoc-members:
    :show-inheritance:
