+++ skbuild/_command/bdist_wheel.py
@@ -16,8 +16,8 @@ else:
         from wheel.bdist_wheel import bdist_wheel as _bdist_wheel
         from wheel.wheelfile import WheelFile
 
+from .._utils import distribution_hide_listing
 from .._version import version as skbuild_version
-from ..utils import distribution_hide_listing
 from . import set_build_base_mixin
