+++ skbuild/_command/sdist.py
@@ -6,7 +6,7 @@ from typing import Sequence
 
 from setuptools.command.sdist import sdist as _sdist
 
-from ..utils import distribution_hide_listing, logger
+from .._utils import distribution_hide_listing, logger
 from . import CommandMixinProtocol, set_build_base_mixin
