+++ skbuild/_platform_specifics/abstract.py
@@ -9,9 +9,9 @@ import subprocess
 import textwrap
 from typing import Iterable, Mapping
 
-from ..constants import CMAKE_DEFAULT_EXECUTABLE
+from .._constants import CMAKE_DEFAULT_EXECUTABLE
+from .._utils import push_dir
 from ..exceptions import SKBuildGeneratorNotFoundError
-from ..utils import push_dir
 
 test_folder = "_cmake_test_compile"
