+++ skbuild/_platform_specifics/cygwin.py
