+++ tests/test_issue342_cmake_osx_args_in_setup.py
@@ -6,7 +6,7 @@ import textwrap
 
 import pytest
 
-import skbuild.constants
+import skbuild._constants
 
 from . import _tmpdir, execute_setup_py, push_env
 
@@ -171,7 +171,7 @@ def test_cmake_args_keyword_osx_default(
     monkeypatch.setattr(sys, "platform", "darwin")
 
     with push_env(MACOSX_DEPLOYMENT_TARGET=osx_deployment_target_env_var):
-        monkeypatch.setattr(skbuild.constants, "_SKBUILD_PLAT_NAME", skbuild.constants._default_skbuild_plat_name())
+        monkeypatch.setattr(skbuild._constants, "_SKBUILD_PLAT_NAME", skbuild._constants._default_skbuild_plat_name())
         with pytest.raises(RuntimeError, match="exit skbuild"):
             with execute_setup_py(tmp_dir, ["build", *cli_setup_args, "--", *cli_cmake_args]):
                 pass
